requirements.txt
python-telegram-bot==20.0
telethon ==1.39.0
requests ==2.32.3
python-dotenv==1.1.0